+++ src/main/java/load/datapool/service/BaseLockerService.java
@@ -22,13 +22,17 @@ public class BaseLockerService implements LockerService {
     private final H2Template jdbcOperations;
     private final TableService tableService;
     private final int batchRows;
+    private final boolean autoLoadPoolFromDB; // For auto load datapool from DB. For debug issues
 
     @Autowired
     public BaseLockerService(H2Template jdbcOperations, TableService tableService,
-                             @Value("${lockerService.batchRows}") int batchRows) {
+                             @Value("${lockerService.batchRows}") int batchRows,
+                             @Value("${lockerService.autoLoadPoolFromDB:false}") boolean autoLoadPoolFromDB
+                             ) {
         this.jdbcOperations = jdbcOperations;
         this.tableService = tableService;
         this.batchRows = batchRows;
+        this.autoLoadPoolFromDB = autoLoadPoolFromDB;
     }
 
     @PostConstruct
@@ -123,7 +127,7 @@ public class BaseLockerService implements LockerService {
         Integer lockedColNum = jdbcOperations.queryForObject(selectColumns, args, Integer.class);
         boolean contain = lockedColNum != null && lockedColNum > 0;
         if (!contain)
-            logger.error("Table " + tableService.fullName(schema, tableName) + "not contain " + column + " column");
+            logger.error("Table " + tableService.fullName(schema, tableName) + " not contain " + column + " column");
         return contain;
     }
 
@@ -143,25 +147,32 @@ public class BaseLockerService implements LockerService {
         if (lockers.get(tableService.fullName(env, pool)) != null){
             return true;
         }
-        final int trueNum = 1;
-        AtomicInteger exist = new AtomicInteger(0);
-        try {
-            showSchemas().forEach(schema -> {
-                if (!schema.equalsIgnoreCase(env))
-                    return;
-                selectTables(schema).forEach(tableName -> {
-                    if (tableName.equalsIgnoreCase(pool))
-                        exist.set(trueNum);
+// При пересоздании пула, возникает коллизия (таблица и сиквенсы еще создаются, а потоки уже пытаются загрузить в приложение этот пул)
+        synchronized (this) {
+            final int trueNum = 1;
+            AtomicInteger exist = new AtomicInteger(0);
+            try {
+                showSchemas().forEach(schema -> {
+                    if (!schema.equalsIgnoreCase(env))
+                        return;
+                    selectTables(schema).forEach(tableName -> {
+                        if (tableName.equalsIgnoreCase(pool))
+                            exist.set(trueNum);
+                    });
                 });
-            });
-        } catch (Exception e) {
-            e.printStackTrace();
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+            if (exist.get() == 1) { //IF there is pool in db, but there is not in lockers
+                if (autoLoadPoolFromDB){
+                    selectLocksFromTable(env, pool); //Load locks
+                    logger.warn("{} was auto loaded from DB. Check DB structure and app lockers cache.", tableService.fullName(env, pool));
+                }
+                logger.warn("{} is presented in database, but is absent in application cache.", tableService.fullName(env, pool));
+            }
+            //return exist.get() == trueNum;
         }
-        if (exist.get() == 1){ //IF there is pool in db, but there is not in lockers
-            selectLocksFromTable(env, pool); //Load locks
-            logger.warn("{} was addition loaded. Check DB structure and app lockers cache.", tableService.fullName(env, pool));
-        }
-        return exist.get() == trueNum;
+        return false;
     }
 
     @Override